more refactor for task runner
